moving types.ts file to utils/types
