+++ utils/transactionExecutor.ts
@@ -1,5 +1,5 @@
 import keyring from '@polkadot/ui-keyring'
-import { KeyringAccount } from '@/types'
+import { KeyringAccount } from '@/utils/types/types'
 import { AddressOrPair, SubmittableExtrinsic } from '@polkadot/api/types'
 import { Callback, ISubmittableResult } from '@polkadot/types/types'
 import { getAddress } from '@/utils/extension'
